% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{clustering}
\alias{clustering}
\title{This function is used to calculate and represent clusters based on the
standard or Euclidean distance}
\usage{
clustering(
  data,
  data_form,
  improve_data,
  numb_cluster,
  elbow_rule,
  optimal_k,
  Graph
)
}
\arguments{
\item{data}{Your dataset with the variables from which you want to obtain clusters by columns}

\item{data_form}{Indicate where are your variables: 'r' is row and default is column}

\item{improve_data}{You have to choice: delete NA values ('NonNA'), sustitute by median values ('Median') or leave as it is (anything difference)}

\item{numb_cluster}{Number of clusters that you want}

\item{elbow_rule}{You can decide whether to use (TRUE) or not (FALSE) the associated elbow rule}

\item{optimal_k}{You can decide whether to use (TRUE) or not (FALSE) the associated optimal elbow rule}

\item{Graph}{You can decide whether to use (TRUE) or not (FALSE) print all associated plots}
}
\value{
Data+clusters, elbow-associated k, optimal elbow-associated k, and plots
}
\description{
This function is used to calculate and represent clusters based on the
standard or Euclidean distance
}
\examples{
# Previously install and read: "mlbench"
data("iris")
data <- iris[,-5]
CK <- clustering(data, 'r', improve_data = FALSE, numb_cluster = 3, elbow_rule = TRUE, optimal_k = TRUE, Graph = TRUE)

}
