% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initial_tests.R
\name{initial_tests}
\alias{initial_tests}
\title{This function give you some basic statistic tests}
\usage{
initial_tests(data, two_by_two = NULL, alpha = NULL)
}
\arguments{
\item{data}{Your dataset with the variables from which you want to analyze}

\item{two_by_two}{Optional param to calculate some test two by two}

\item{alpha}{Optional param to change alpha for all tests which have that option}
}
\value{
t.test, shapiro.test, var.test, kurtosis, skrewness
}
\description{
This function give you some basic statistic tests
}
\examples{
# Previously install and read: "mlbench"
data("iris")
data <- iris[,-5]
test <- initial_tests(data, two_by_two = TRUE)

}
