% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/susmedian.R
\name{susmedian}
\alias{susmedian}
\title{This function is used to improve your NAs data by a median value of its measure}
\usage{
susmedian(data, data_form)
}
\arguments{
\item{data}{Your data set}

\item{data_form}{Indicate where are your variables: 'r' is row and default is
column but you have to write something}
}
\value{
Data improve by median values
}
\description{
This function is used to improve your NAs data by a median value of its measure
}
\examples{
matriz <- matrix(c(1, NA, 3, 4, 5, 6, NA, 8, 9), nrow = 3, ncol = 3, byrow = TRUE)
colnames(matriz) <- c("Variable1", "Variable2", "Variable3")
susmedian(matriz,data_form = 'c')
 
}
